+++ cognite/client/_api/simulators/__init__.py
@@ -4,6 +4,7 @@ from collections.abc import Iterator
 from typing import TYPE_CHECKING, overload
 
 from cognite.client._api.simulators.integrations import SimulatorIntegrationsAPI
+from cognite.client._api.simulators.logs import SimulatorLogsAPI
 from cognite.client._api.simulators.models import SimulatorModelsAPI
 from cognite.client._api.simulators.routines import SimulatorRoutinesAPI
 from cognite.client._api_client import APIClient
@@ -24,6 +25,7 @@ class SimulatorsAPI(APIClient):
         self.integrations = SimulatorIntegrationsAPI(config, api_version, cognite_client)
         self.models = SimulatorModelsAPI(config, api_version, cognite_client)
         self.routines = SimulatorRoutinesAPI(config, api_version, cognite_client)
+        self.logs = SimulatorLogsAPI(config, api_version, cognite_client)
         self._warning = FeaturePreviewWarning(
             api_maturity="General Availability", sdk_maturity="alpha", feature_name="Simulators"
         )