+++ tests/tests_integration/test_api/test_simulators/test_integrations.py
@@ -36,6 +36,14 @@ class TestSimulatorIntegrations:
         assert item.active is True
         assert item.created_time is not None
         assert item.last_updated_time is not None
+        assert item.log_id is not None
+        log = cognite_client.simulators.logs.retrieve(id=item.log_id)
+        assert log is not None
+        assert log.data is not None
+        assert log.data[0].timestamp is not None
+        assert log.data[0].timestamp > int(time.time() * 1000) - 10000  # updated less than 10 seconds ago
+        assert log.data[0].message == "Testing logs update for simulator integration"
+        assert log.data[0].severity == "Debug"
 
     def test_delete_integrations(self, cognite_client: CogniteClient, seed_resource_names) -> None:
         simulator_integration["heartbeat"] = int(time.time() * 1000)